Refactor primitive keys to their own package
