--- src/main/java/com/trivago/fastutilconcurrentwrapper/PrimitiveIntKeyMap.java
@@ -1,5 +1,0 @@
-package com.trivago.fastutilconcurrentwrapper;
-
-public interface PrimitiveIntKeyMap extends KeyMap {
-    boolean containsKey(int key);
-}
