--- src/main/java/com/trivago/fastutilconcurrentwrapper/PrimitiveLongKeyMap.java
@@ -1,5 +1,0 @@
-package com.trivago.fastutilconcurrentwrapper;
-
-public interface PrimitiveLongKeyMap extends KeyMap {
-    boolean containsKey(long key);
-}
