+++ src/main/java/com/trivago/fastutilconcurrentwrapper/primitivekeys/LongIntMap.java
@@ -1,4 +1,4 @@
-package com.trivago.fastutilconcurrentwrapper;
+package com.trivago.fastutilconcurrentwrapper.primitivekeys;
 
 import it.unimi.dsi.fastutil.longs.Long2IntFunction;
