+++ src/main/java/com/trivago/fastutilconcurrentwrapper/primitivekeys/PrimitiveIntKeyMap.java
@@ -1,0 +1,7 @@
+package com.trivago.fastutilconcurrentwrapper.primitivekeys;
+
+import com.trivago.fastutilconcurrentwrapper.KeyMap;
+
+public interface PrimitiveIntKeyMap extends KeyMap {
+    boolean containsKey(int key);
+}
