+++ src/main/java/com/trivago/fastutilconcurrentwrapper/primitivekeys/PrimitiveLongKeyMap.java
@@ -1,0 +1,7 @@
+package com.trivago.fastutilconcurrentwrapper.primitivekeys;
+
+import com.trivago.fastutilconcurrentwrapper.KeyMap;
+
+public interface PrimitiveLongKeyMap extends KeyMap {
+    boolean containsKey(long key);
+}
