+++ src/main/java/com/trivago/fastutilconcurrentwrapper/primitivekeys/map/ConcurrentIntIntMap.java
@@ -1,28 +1,30 @@
-package com.trivago.fastutilconcurrentwrapper.map;
+package com.trivago.fastutilconcurrentwrapper.primitivekeys.map;
 
-import com.trivago.fastutilconcurrentwrapper.IntFloatMap;
-import com.trivago.fastutilconcurrentwrapper.wrapper.PrimitiveFastutilIntFloatWrapper;
-import it.unimi.dsi.fastutil.ints.Int2FloatFunction;
+import com.trivago.fastutilconcurrentwrapper.primitivekeys.IntIntMap;
+import com.trivago.fastutilconcurrentwrapper.primitivekeys.wrapper.PrimitiveFastutilIntIntWrapper;
+import it.unimi.dsi.fastutil.ints.Int2IntFunction;
 
 import java.util.concurrent.locks.Lock;
 import java.util.function.BiFunction;
 
-public class ConcurrentIntFloatMap extends PrimitiveConcurrentMap implements IntFloatMap {
+public class ConcurrentIntIntMap extends PrimitiveConcurrentMap implements IntIntMap {
 
-    private final IntFloatMap[] maps;
-    private final float defaultValue;
+    private final IntIntMap[] maps;
+    private final int defaultValue;
 
-    public ConcurrentIntFloatMap(int numBuckets,
-                                 int initialCapacity,
-                                 float loadFactor,
-                                 float defaultValue) {
+    public ConcurrentIntIntMap(
+            int numBuckets,
+            int initialCapacity,
+            float loadFactor,
+            int defaultValue) {
+
         super(numBuckets);
 
-        this.maps = new IntFloatMap[numBuckets];
+        this.maps = new IntIntMap[numBuckets];
         this.defaultValue = defaultValue;
 
         for (int i = 0; i < numBuckets; i++) {
-            maps[i] = new PrimitiveFastutilIntFloatWrapper(initialCapacity, loadFactor, defaultValue);
+            maps[i] = new PrimitiveFastutilIntIntWrapper(initialCapacity, loadFactor, defaultValue);
         }
     }
 
@@ -50,10 +52,10 @@ public class ConcurrentIntFloatMap extends PrimitiveConcurrentMap implements Int
     }
 
     @Override
-    public float get(int l) {
+    public int get(int l) {
         int bucket = getBucket(l);
 
-        float result;
+        int result;
 
         Lock readLock = locks[bucket].readLock();
         readLock.lock();
@@ -67,10 +69,10 @@ public class ConcurrentIntFloatMap extends PrimitiveConcurrentMap implements Int
     }
 
     @Override
-    public float put(int key, float value) {
+    public int put(int key, int value) {
         int bucket = getBucket(key);
 
-        float result;
+        int result;
 
         Lock writeLock = locks[bucket].writeLock();
         writeLock.lock();
@@ -84,12 +86,12 @@ public class ConcurrentIntFloatMap extends PrimitiveConcurrentMap implements Int
     }
 
     @Override
-    public float getDefaultValue() {
+    public int getDefaultValue() {
         return defaultValue;
     }
 
     @Override
-    public float remove(int key) {
+    public int remove(int key) {
         int bucket = getBucket(key);
 
         Lock writeLock = locks[bucket].writeLock();
@@ -102,7 +104,7 @@ public class ConcurrentIntFloatMap extends PrimitiveConcurrentMap implements Int
     }
 
     @Override
-    public boolean remove(int key, float value) {
+    public boolean remove(int key, int value) {
         int bucket = getBucket(key);
 
         Lock writeLock = locks[bucket].writeLock();
@@ -115,7 +117,7 @@ public class ConcurrentIntFloatMap extends PrimitiveConcurrentMap implements Int
     }
 
     @Override
-    public float computeIfAbsent(int key, Int2FloatFunction mappingFunction) {
+    public int computeIfAbsent(int key, Int2IntFunction mappingFunction) {
         int bucket = getBucket(key);
 
         Lock writeLock = locks[bucket].writeLock();
@@ -128,7 +130,7 @@ public class ConcurrentIntFloatMap extends PrimitiveConcurrentMap implements Int
     }
 
     @Override
-    public float computeIfPresent(int key, BiFunction<Integer, Float, Float> mappingFunction) {
+    public int computeIfPresent(int key, BiFunction<Integer, Integer, Integer> mappingFunction) {
         int bucket = getBucket(key);
 
         Lock writeLock = locks[bucket].writeLock();
