+++ src/main/java/com/trivago/fastutilconcurrentwrapper/primitivekeys/map/ConcurrentLongLongMap.java
@@ -1,33 +1,28 @@
-package com.trivago.fastutilconcurrentwrapper.map;
+package com.trivago.fastutilconcurrentwrapper.primitivekeys.map;
 
-import com.trivago.fastutilconcurrentwrapper.IntIntMap;
-import com.trivago.fastutilconcurrentwrapper.LongIntMap;
-import com.trivago.fastutilconcurrentwrapper.wrapper.PrimitiveFastutilIntIntWrapper;
-import com.trivago.fastutilconcurrentwrapper.wrapper.PrimitiveFastutilLongIntWrapper;
-import it.unimi.dsi.fastutil.ints.Int2IntFunction;
-import it.unimi.dsi.fastutil.longs.Long2IntFunction;
+import com.trivago.fastutilconcurrentwrapper.primitivekeys.LongLongMap;
+import com.trivago.fastutilconcurrentwrapper.primitivekeys.wrapper.PrimitiveFastutilLongLongWrapper;
+import it.unimi.dsi.fastutil.longs.Long2LongFunction;
 
 import java.util.concurrent.locks.Lock;
 import java.util.function.BiFunction;
 
-public class ConcurrentLongIntMap extends PrimitiveConcurrentMap implements LongIntMap {
+public class ConcurrentLongLongMap extends PrimitiveConcurrentMap implements LongLongMap {
 
-    private final LongIntMap[] maps;
-    private final int defaultValue;
+    private final LongLongMap[] maps;
+    private final long defaultValue;
 
-    public ConcurrentLongIntMap(
-            int numBuckets,
-            int initialCapacity,
-            float loadFactor,
-            int defaultValue) {
-
+    public ConcurrentLongLongMap(int numBuckets,
+                                 int initialCapacity,
+                                 float loadFactor,
+                                 long defaultValue) {
         super(numBuckets);
 
-        this.maps = new LongIntMap[numBuckets];
+        this.maps = new LongLongMap[numBuckets];
         this.defaultValue = defaultValue;
 
         for (int i = 0; i < numBuckets; i++) {
-            maps[i] = new PrimitiveFastutilLongIntWrapper(initialCapacity, loadFactor, defaultValue);
+            maps[i] = new PrimitiveFastutilLongLongWrapper(initialCapacity, loadFactor, defaultValue);
         }
     }
 
@@ -55,10 +50,10 @@ public class ConcurrentLongIntMap extends PrimitiveConcurrentMap implements Long
     }
 
     @Override
-    public int get(long l) {
+    public long get(long l) {
         int bucket = getBucket(l);
 
-        int result;
+        long result;
 
         Lock readLock = locks[bucket].readLock();
         readLock.lock();
@@ -72,10 +67,10 @@ public class ConcurrentLongIntMap extends PrimitiveConcurrentMap implements Long
     }
 
     @Override
-    public int put(long key, int value) {
+    public long put(long key, long value) {
         int bucket = getBucket(key);
 
-        int result;
+        long result;
 
         Lock writeLock = locks[bucket].writeLock();
         writeLock.lock();
@@ -89,12 +84,12 @@ public class ConcurrentLongIntMap extends PrimitiveConcurrentMap implements Long
     }
 
     @Override
-    public int getDefaultValue() {
+    public long getDefaultValue() {
         return defaultValue;
     }
 
     @Override
-    public int remove(long key) {
+    public long remove(long key) {
         int bucket = getBucket(key);
 
         Lock writeLock = locks[bucket].writeLock();
@@ -107,7 +102,7 @@ public class ConcurrentLongIntMap extends PrimitiveConcurrentMap implements Long
     }
 
     @Override
-    public boolean remove(long key, int value) {
+    public boolean remove(long key, long value) {
         int bucket = getBucket(key);
 
         Lock writeLock = locks[bucket].writeLock();
@@ -120,7 +115,7 @@ public class ConcurrentLongIntMap extends PrimitiveConcurrentMap implements Long
     }
 
     @Override
-    public int computeIfAbsent(long key, Long2IntFunction mappingFunction) {
+    public long computeIfAbsent(long key, Long2LongFunction mappingFunction) {
         int bucket = getBucket(key);
 
         Lock writeLock = locks[bucket].writeLock();
@@ -133,7 +128,7 @@ public class ConcurrentLongIntMap extends PrimitiveConcurrentMap implements Long
     }
 
     @Override
-    public int computeIfPresent(long key, BiFunction<Long, Integer, Integer> mappingFunction) {
+    public long computeIfPresent(long key, BiFunction<Long, Long, Long> mappingFunction) {
         int bucket = getBucket(key);
 
         Lock writeLock = locks[bucket].writeLock();
